@@ -48,40 +48,30 @@ dependencies = [
     # LLM Integration
     "langchain-groq>=1.0.0",
     "langchain-openai>=1.0.0",
-
     # Data Science & ML
     "scipy>=1.14.1",
     "numpy>=1.24.0",
     "pandas>=2.0.0",
-    "scikit-learn>=1.3.0",
-
+    "scikit-learn>=1.3.0,<1.8",
     # NLP & Embeddings
     "nltk>=3.8.0",
     "transformers>=4.35.0",
     "sentence-transformers>=5.0.0",
-
-    # PyTorch (required by transformers & sentence-transformers)
     "torch>=2.0.0",
-
-    # Clustering & Dimensionality Reduction
-    "umap-learn>=0.5.0",
-    "hdbscan>=0.8.0",
-
     # Data Validation & Utilities
     "pydantic>=2.0.0,<3.0.0",
     "tqdm>=4.65.0",
     "jinja2>=3.1.0",
+    "umap-learn>=0.5.9.post2",
+    "hdbscan>=0.8.41",
 ]
 
-# Optional dependencies for specific features
 [project.optional-dependencies]
-# Cloud storage support (AWS S3, LakeFS)
 cloud = [
     "boto3>=1.28.0",
     "lakefs>=0.14.0",
 ]
 
-# Full installation with all features
 all = [
     "alquimia-fair-forge[cloud]",
 ]
@@ -170,7 +160,7 @@ exclude = [
 # Ruff configuration - Modern, fast Python linter & formatter
 [tool.ruff]
 line-length = 120
-target-version = "py310"
+target-version = "py313"
 extend-exclude = [
     ".git",
     ".venv",
@@ -236,7 +226,7 @@ line-ending = "auto"
 
 # MyPy configuration - Static type checking
 [tool.mypy]
-python_version = "3.10"
+python_version = "3.13"
 warn_return_any = true
 warn_unused_configs = true
 warn_redundant_casts = true
@@ -296,10 +286,12 @@ markers = [
     "requires_api_key: marks tests that require API keys",
 ]
 filterwarnings = [
-    "error",
     "ignore::DeprecationWarning",
     "ignore::PendingDeprecationWarning",
     "ignore::UserWarning:pkg_resources",
+    "ignore::SyntaxWarning",  
+    "ignore::ImportWarning:umap",  
+    "error",  
 ]
 
 # Coverage configuration
